Fix pipe-related deadlock
If one of the programs run by init produces a large amount of stderr or
stdout, the pipe buffer in kernelspace will eventually fill up, causing
the launched program to block in a write. Init likewise will be blocked
on Wait4, waiting for the launched program to finish, resulting in a
deadlock.

This commit spins up two goroutines to constanly drain data from stdout
and stderr so this doesn't happen.